requirements.txt
asgiref==3.8.1
boto3==1.36.26
botocore==1.36.26
coverage==7.6.12
Django==5.1.5
django-appconf==1.1.0
django-autocomplete-light==3.11.0
django-magiclink==1.3.0
django-select2==8.3.0
djhacker==0.2.3
gunicorn==20.1.0
iniconfig==2.0.0
jmespath==1.0.1
packaging==21.3
pluggy==1.5.0
pyparsing==3.2.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
